[Observability] Relocate MP observability to lmcache/v1/mp_observability (#2657)
* move the mp mode observability to lmcache/v1/mp_observability


* skip the cpu test


---------
