+++ .github/workflows/test.yml
@@ -83,6 +83,7 @@ jobs:
           pytest --ignore=tests/disagg --ignore=tests/v1/test_nixl_storage.py \
             --ignore=tests/v1/multiprocess/ \
             --ignore=tests/v1/distributed/ \
+            --ignore=tests/v1/mp_observability/ \
             --ignore=tests/skipped \
             --ignore=tests/v1/storage_backend/test_eic.py
